Updated a lot of includes, and the Makefile a bit. Still needs work. Oh, but the binGT function is awesome now. I dunno why I never thought to XOR before.

+++ architectures/16.h
@@ -1,5 +1,2 @@
-#define BIN_INT uint
-#define BIN_BITS sizeof(BIN_INT)*4
-
-
-#include "../bin/bin.h"
+#define BIN_INT int
+#define BIN_BITS (int)sizeof(BIN_INT)*4
